+++ lib/view_component/template.rb
@@ -5,56 +5,112 @@ module ViewComponent
     DataWithSource = Struct.new(:format, :identifier, :short_identifier, :type, keyword_init: true)
     DataNoSource = Struct.new(:source, :identifier, :type, keyword_init: true)
 
-    attr_reader :variant, :this_format, :type
+    attr_reader :details
+
+    delegate :format, :variant, to: :details
 
     def initialize(
       component:,
-      type:,
-      this_format: nil,
-      variant: nil,
+      details:,
       lineno: nil,
       path: nil,
-      extension: nil,
-      source: nil,
-      method_name: nil,
-      defined_on_self: true
+      method_name: nil
     )
       @component = component
-      @type = type
-      @this_format = this_format
-      @variant = variant&.to_sym
+      @details = details
       @lineno = lineno
       @path = path
-      @extension = extension
-      @source = source
       @method_name = method_name
-      @defined_on_self = defined_on_self
-
-      @source_originally_nil = @source.nil?
 
       @call_method_name =
         if @method_name
           @method_name
         else
           out = +"call"
-          out << "_#{normalized_variant_name}" if @variant.present?
-          out << "_#{@this_format}" if @this_format.present? && @this_format != ViewComponent::Base::VC_INTERNAL_DEFAULT_FORMAT
+          out << "_#{normalized_variant_name}" if variant.present?
+          out << "_#{format}" if format.present? && format != ViewComponent::Base::VC_INTERNAL_DEFAULT_FORMAT
           out
         end
     end
 
-    def compile_to_component
-      if !inline_call?
-        @component.silence_redefinition_of_method(@call_method_name)
-
-        # rubocop:disable Style/EvalWithLocation
-        @component.class_eval <<-RUBY, @path, @lineno
-        def #{@call_method_name}
-          #{compiled_source}
-        end
-        RUBY
-        # rubocop:enable Style/EvalWithLocation
+    class File < Template
+      def initialize(component:, details:, path:)
+        super(
+          component: component,
+          details: details,
+          path: path,
+          lineno: 0
+        )
       end
+
+      def type
+        :file
+      end
+
+      # Load file each time we look up #source in case the file has been modified
+      def source
+        ::File.read(@path)
+      end
+    end
+
+    class Inline < Template
+      attr_reader :source
+
+      def initialize(component:, inline_template:)
+        details = ActionView::TemplateDetails.new(nil, inline_template.language.to_sym, nil, nil)
+
+        super(
+          component: component,
+          details: details,
+          path: inline_template.path,
+          lineno: inline_template.lineno,
+        )
+
+        @source = inline_template.source.dup
+      end
+
+      def type
+        :inline
+      end
+    end
+
+    class InlineCall < Template
+      def initialize(component:, method_name:, defined_on_self:)
+        variant = method_name.to_s.include?("call_") ? method_name.to_s.sub("call_", "").to_sym : nil
+        details = ActionView::TemplateDetails.new(nil, nil, nil, variant)
+
+        super(
+          component: component,
+          details: details,
+          method_name: method_name
+        )
+
+        @defined_on_self = defined_on_self
+      end
+
+      def type
+        :inline_call
+      end
+
+      def compile_to_component
+        @component.define_method(safe_method_name, @component.instance_method(@call_method_name))
+      end
+
+      def defined_on_self?
+        @defined_on_self
+      end
+    end
+
+    def compile_to_component
+      @component.silence_redefinition_of_method(@call_method_name)
+
+      # rubocop:disable Style/EvalWithLocation
+      @component.class_eval <<-RUBY, @path, @lineno
+      def #{@call_method_name}
+        #{compiled_source}
+      end
+      RUBY
+      # rubocop:enable Style/EvalWithLocation
 
       @component.define_method(safe_method_name, @component.instance_method(@call_method_name))
     end
@@ -72,19 +128,15 @@ module ViewComponent
     end
 
     def inline_call?
-      @type == :inline_call
+      type == :inline_call
     end
 
     def inline?
-      @type == :inline
+      type == :inline
     end
 
     def default_format?
-      @this_format == ViewComponent::Base::VC_INTERNAL_DEFAULT_FORMAT
-    end
-
-    def format
-      @this_format
+      format.nil? || format == ViewComponent::Base::VC_INTERNAL_DEFAULT_FORMAT
     end
 
     def safe_method_name
@@ -92,35 +144,23 @@ module ViewComponent
     end
 
     def normalized_variant_name
-      @variant.to_s.gsub("-", "__").gsub(".", "___")
-    end
-
-    def defined_on_self?
-      @defined_on_self
+      variant.to_s.gsub("-", "__")
     end
 
     private
 
-    def source
-      if @source_originally_nil
-        # Load file each time we look up #source in case the file has been modified
-        File.read(@path)
-      else
-        @source
-      end
-    end
-
     def compiled_source
-      handler = ActionView::Template.handler_for_extension(@extension)
+      handler = details.handler_class
       this_source = source
       this_source.rstrip! if @component.strip_trailing_whitespace?
 
       short_identifier = defined?(Rails.root) ? @path.sub("#{Rails.root}/", "") : @path
-      type = ActionView::Template::Types[@this_format]
+      format = self.format || ViewComponent::Base::VC_INTERNAL_DEFAULT_FORMAT
+      type = ActionView::Template::Types[format]
 
       if handler.method(:call).parameters.length > 1
         handler.call(
-          DataWithSource.new(format: @this_format, identifier: @path, short_identifier: short_identifier, type: type),
+          DataWithSource.new(format: format, identifier: @path, short_identifier: short_identifier, type: type),
           this_source
         )
       # :nocov: