+++ docs/CHANGELOG.md
@@ -18,6 +18,10 @@ nav_order: 5
 
     *Joel Hawksley*
 
+* Add CI coverage for Rails 8.
+
+    *Joel Hawksley*
+
 ## 3.20.0
 
 * Allow rendering `with_collection` to accept an optional `spacer_component` to be rendered between each item.